feat(269): client-side transformations; move transformData to utils
